requirements.txt
asn1crypto==0.22.0
bcrypt==3.1.3
Beaker==1.5.4
beautifulsoup4==4.6.0
certifi==2017.7.27.1
cffi==1.9.1
chardet==3.0.4
configparser==3.5.0
coverage==4.4.1
cryptography==2.0.2
cssselect==0.9.2
decorator==4.0.11
enum-compat==0.0.2
enum34==1.1.6
eventlet==0.21.0
greenlet==0.4.12
gssapi==1.2.0
html5lib==0.999
humanize==0.5.1
hurry.filesize==0.9
idna==2.6
iniparse==0.4
ipaddress==1.0.18
IPy==0.81
mysqlclient==1.3.12
paramiko==2.2.1
pbr==3.1.1
ply==3.9
psutil==5.0.1
pyasn1==0.2.3
pycodestyle==2.3.1
pycparser==2.18
pycrypto==2.6.1
pymongo==3.5.1
pymongo-amplidata==2.8.post2
PyNaCl==1.1.1
pyodbc==4.0.17
requests==2.18.4
six==1.11.0
stevedore==1.26.0
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.8.2
